linux: Use float math functions

+++ src/core/vec.cpp
@@ -135,15 +135,15 @@ operator!=(fvec3 a, fvec3 b) noexcept {
 
 float
 distanceTo(fvec2 a, fvec2 b) noexcept {
-    double dx = static_cast<double>(a.x) - static_cast<double>(b.x);
-    double dy = static_cast<double>(a.y) - static_cast<double>(b.y);
-    return static_cast<float>(sqrt(dx * dx + dy * dy));
+    float dx = a.x - b.x;
+    float dy = a.y - b.y;
+    return sqrtf(dx * dx + dy * dy);
 }
 float
 distanceTo(fvec3 a, fvec3 b) noexcept {
-    double dx = static_cast<double>(a.x) - static_cast<double>(b.x);
-    double dy = static_cast<double>(a.y) - static_cast<double>(b.y);
-    return static_cast<float>(sqrt(dx * dx + dy * dy));
+    float dx = a.x - b.x;
+    float dy = a.y - b.y;
+    return sqrtf(dx * dx + dy * dy);
 }
 
 size_t
